+++ ner_telhub/model/file_models.py
@@ -1,12 +1,10 @@
 import os
 from typing import List
-
 from PyQt6.QtWidgets import QWidget
 from PyQt6.QtCore import (
     QAbstractListModel, Qt,
     pyqtBoundSignal, QModelIndex,
 )
-
 from ner_processing.decode_files import LogFormat, processLine
 from ner_processing.message import Message
 from ner_telhub.model.data_models import DataModelManager
