+++ ner_telhub/model/filter_models.py
@@ -1,11 +1,9 @@
 from typing import List, Any, Tuple
-
 from PyQt6.QtWidgets import QWidget
 from PyQt6.QtCore import (
     QAbstractListModel, Qt,
     QModelIndex
 )
-
 from ner_telhub.model.message_models import MessageModel
