+++ ner_telhub/view/sd_card/window.py
@@ -1,468 +1,17 @@
 from typing import Dict
 from PyQt6.QtWidgets import (
-    QMainWindow, QLabel, QHBoxLayout,
-    QVBoxLayout, QWidget, QFileDialog,
-    QListView, QTextEdit,
-    QGridLayout, QProgressBar, QDialog,
-    QDialogButtonBox, QLineEdit, QMessageBox,
-    QCheckBox, QMenu
+    QMainWindow, QHBoxLayout,
+    QVBoxLayout, QWidget, QMenu
 )
 from PyQt6.QtGui import QAction
-from PyQt6.QtCore import QSize, Qt
-
-from ner_telhub import colors
 from ner_telhub.model.file_models import FileModel
 from ner_telhub.model.data_models import DataModelManager
 from ner_processing.decode_files import LogFormat
+from ner_telhub.view.sd_card.file_view import FileView
+from ner_telhub.view.sd_card.options_view import OptionsView
+from ner_telhub.view.sd_card.process_view import ProcessView
 from ner_telhub.widgets.menu_widgets import MessageIds, DataIds
-from ner_telhub.widgets.graphing_widgets import GraphDashboardWidget
-from ner_telhub.widgets.styled_widgets import (
-    NERButton, NERToolbar,
-    NERLoadingSpinner, NERInfoButton
-)
-from ner_telhub.utils.timestream import TimestreamIngestionService
-from ner_telhub.descriptions import SD_OPTIONS_DESC
-
-
-class GraphDialog(QDialog):
-    """Shows a dashboard for data in the system."""
-
-    def __init__(self, parent: QWidget, model: DataModelManager):
-        super().__init__(parent)
-        self.setWindowTitle("Graph View")
-
-        toolbar = NERToolbar()
-        self.setStyleSheet(
-            "QToolBar { background-color: " +
-            colors.SECONDARY_BACKGROUND +
-            "; padding: 5%; border: none}")
-        self.screen_button = NERButton("Full Screen", NERButton.Styles.GRAY)
-        self.screen_button.pressed.connect(self.changeScreen)
-        toolbar.addLeft(self.screen_button)
-
-        layout = QVBoxLayout()
-        layout.addWidget(toolbar)
-        layout.addWidget(GraphDashboardWidget(parent, model, False))
-        self.setLayout(layout)
-
-    def changeScreen(self):
-        if self.screen_button.text() == "Full Screen":
-            self.showFullScreen()
-            self.screen_button.setText("Minimize")
-        else:
-            self.showNormal()
-            self.screen_button.setText("Full Screen")
-
-
-class ExportDialog(QDialog):
-    """Dialog to export data to a CSV file."""
-
-    def __init__(
-            self,
-            parent: QWidget,
-            model: DataModelManager,
-            spinner: NERLoadingSpinner):
-        super().__init__(parent)
-
-        self.setWindowTitle("Export CSV")
-        self.model = model
-        self.spinner = spinner
-
-        self.filename_input = QLineEdit()
-        self.directory_input = QLineEdit()
-        self.directory_button = NERButton("...", NERButton.Styles.GRAY)
-        self.directory_button.addStyle("font-size: 20px")
-        self.directory_button.setMaximumSize(45, 25)
-        self.directory_button.pressed.connect(self.choose_directory)
-        self.directory_button.setToolTip("Choose from file system")
-
-        self.layout = QGridLayout()
-        self.layout.addWidget(QLabel("File name: "), 0, 0)
-        self.layout.addWidget(self.filename_input, 0, 1)
-        self.layout.addWidget(QLabel("Directory name: "), 1, 0)
-        self.layout.addWidget(self.directory_input, 1, 1)
-        self.layout.addWidget(self.directory_button, 1, 2)
-
-        self.buttonBox = QDialogButtonBox(
-            QDialogButtonBox.StandardButton.Ok | QDialogButtonBox.StandardButton.Cancel)
-        self.buttonBox.accepted.connect(self.on_accept)
-        self.buttonBox.rejected.connect(self.reject)
-
-        self.layout.addWidget(self.buttonBox)
-        self.setLayout(self.layout)
-
-    def choose_directory(self):
-        dir = QFileDialog().getExistingDirectory(self)
-        self.directory_input.setText(dir)
-
-    def on_accept(self):
-        try:
-            filename = self.create_extension(self.filename_input.text())
-        except BaseException:
-            QMessageBox.critical(
-                self,
-                "Invalid File Name",
-                "File must be either a .csv or contain no extension.")
-            return
-
-        directory = self.directory_input.text()
-        full_path = directory + "/" + filename
-
-        worker = self.model.getCSVWorker(full_path)
-        worker.signals.finished.connect(lambda: self.spinner.stopAnimation())
-        worker.signals.error.connect(
-            lambda error: QMessageBox.critical(
-                self, "Export Error", error[1].__str__()))
-        worker.signals.message.connect(
-            lambda msg: QMessageBox.about(
-                self, "Export Status", msg))
-        try:
-            worker.start()
-            self.spinner.startAnimation()
-        except RuntimeError as e:
-            QMessageBox.critical(self, "Internal Error", str(e))
-        self.accept()
-
-    @staticmethod
-    def create_extension(name: str):
-        """Verifies the file name has a csv extension, or adds one if not."""
-        components = name.split(".")
-        if len(components) == 1:
-            return name + ".csv"
-        elif len(components) == 2 and components[1] == "csv":
-            return name
-        else:
-            raise ValueError("Invalid file name format")
-
-
-class DatabaseDialog(QDialog):
-    """Dialog to export data to a the database."""
-
-    def __init__(
-            self,
-            parent: QWidget,
-            model: DataModelManager,
-            spinner: NERLoadingSpinner):
-        super().__init__(parent)
-
-        self.setWindowTitle("Database Export")
-        self.model = model
-        self.spinner = spinner
-
-        self.testid_input = QLineEdit()
-        self.label = QLabel(
-            """The test ID is a unique identifier that you can use to identify a testing session.
-Make sure the test ID is a meaningful name so you can reload previous sessions by their ID.""")
-
-        self.layout = QVBoxLayout()
-        self.layout.addWidget(QLabel("Test ID:"))
-        self.layout.addWidget(self.testid_input)
-        self.layout.addWidget(self.label)
-
-        self.buttonBox = QDialogButtonBox(
-            QDialogButtonBox.StandardButton.Ok | QDialogButtonBox.StandardButton.Cancel)
-        self.buttonBox.accepted.connect(self.on_accept)
-        self.buttonBox.rejected.connect(self.reject)
-
-        self.layout.addWidget(self.buttonBox)
-        self.setLayout(self.layout)
-
-    def on_accept(self):
-        if self.model.getDataCount() == 0:
-            QMessageBox.critical(
-                self, "Error", "Cannot export data - data model is empty.")
-            return
-
-        testid = self.testid_input.text()
-        if testid == "":
-            QMessageBox.critical(self, "Error", "You must specify a test ID")
-            return
-
-        button = QMessageBox.question(
-            self,
-            "Warning",
-            "You cannot undo this action, so please make sure you are authorized to complete it."
-            "\nWould you like to continue?")
-        if button == QMessageBox.StandardButton.No:
-            return
-
-        models = [self.model.getDataModel(id)
-                  for id in self.model.getAvailableIds()]
-        worker = TimestreamIngestionService.getIngestionWorker(models, testid)
-        worker.signals.finished.connect(lambda: self.spinner.stopAnimation())
-        worker.signals.error.connect(
-            lambda error: QMessageBox.critical(
-                self, "Export Error", error[1].__str__()))
-        worker.signals.message.connect(
-            lambda msg: QMessageBox.about(
-                self, "Export Status", msg))
-        try:
-            worker.start()
-            self.spinner.startAnimation()
-        except RuntimeError as e:
-            QMessageBox.critical(self, "Internal Error", str(e))
-        self.accept()
-
-
-class FileView(QWidget):
-    """View section with file information and control."""
-
-    def __init__(self, parent: QWidget, file_model: FileModel):
-        super(FileView, self).__init__(parent)
-
-        self.file_model = file_model
-
-        header = QLabel("Selected Files")
-        header.setStyleSheet("font-size: 16px; font-weight: 400")
-        header.setAlignment(Qt.AlignmentFlag.AlignHCenter)
-
-        self.view = QListView()
-        self.view.setModel(self.file_model)
-
-        self.add_button = NERButton("Add Files", NERButton.Styles.GREEN)
-        self.add_button.setToolTip("Add files from the file system to process")
-        self.remove_button = NERButton("Remove Files", NERButton.Styles.RED)
-        self.remove_button.setToolTip("Remove file(s) from the text box")
-        self.add_button.pressed.connect(self.add_files)
-        self.remove_button.pressed.connect(self.remove_files)
-
-        layout = QVBoxLayout()
-        layout.addWidget(header)
-        layout.addWidget(self.view)
-        layout.addWidget(self.add_button)
-        layout.addWidget(self.remove_button)
-
-        self.setLayout(layout)
-
-    def add_files(self):
-        filepaths = QFileDialog().getOpenFileNames()[0]
-        for fp in filepaths:
-            self.file_model.addFile(fp)
-
-    def remove_files(self):
-        indexes = self.view.selectedIndexes()
-        if indexes:
-            index = indexes[0]
-            self.file_model.deleteFile(index)
-            self.view.clearSelection()
-        else:
-            self.file_model.deleteAll()
-
-
-class ProcessView(QWidget):
-    """View section with processing information an control."""
-
-    def __init__(
-            self,
-            parent: QWidget,
-            file_model: FileModel,
-            data_model: DataModelManager):
-        super(ProcessView, self).__init__(parent)
-
-        self.file_model = file_model
-        self.data_model = data_model
-
-        header = QLabel("Processing Information")
-        header.setStyleSheet("font-size: 16px; font-weight: 400")
-        header.setAlignment(Qt.AlignmentFlag.AlignHCenter)
-
-        self.view = QTextEdit()
-        self.view.setReadOnly(True)
-        self.view_text = ""
-
-        self.start_button = NERButton("Start", NERButton.Styles.GREEN)
-        self.start_button.pressed.connect(self.start_process)
-        self.start_button.setToolTip(
-            "Start/stop processing the specified log files")
-
-        # Setup Progress Bar
-        self.progress_bar = QProgressBar()
-        self.progress_bar.setVisible(False)
-        self.progress_bar.setMinimum(0)
-        self.progress_bar.setMaximum(100)
-
-        layout = QVBoxLayout()
-        layout.addWidget(header)
-        layout.addWidget(self.view)
-        layout.addWidget(self.start_button)
-        layout.addWidget(self.progress_bar)
-        self.setLayout(layout)
-
-    def start_process(self):
-        if not self.data_model.isEmpty():
-            button = QMessageBox.question(
-                self, "Warning", "Current model data will be lost. \n"
-                "Would you like to continue?")
-            if button == QMessageBox.StandardButton.No:
-                return
-
-        self.data_model.deleteAllData()
-        self.start_button.setEnabled(False)
-        self.progress_bar.setVisible(True)
-        self.clear_view()
-
-        worker = self.file_model.getProcessWorker(self.data_model)
-        worker.signals.finished.connect(self.stop_process)
-        worker.signals.error.connect(
-            lambda error: QMessageBox.critical(
-                self, "Processing Error", error[1].__str__()))
-        worker.signals.message.connect(self.update_view)
-        worker.signals.progress.connect(
-            lambda val: self.progress_bar.setValue(val))
-
-        try:
-            worker.start()
-        except RuntimeError as e:
-            self.stop_process()
-            QMessageBox.critical(self, "Internal Error", str(e))
-
-    def stop_process(self):
-        self.start_button.setEnabled(True)
-        self.progress_bar.setVisible(False)
-        self.progress_bar.setValue(0)
-
-    def update_view(self, message: str):
-        self.view_text += message
-        self.view_text += "\n"
-        self.view.setText(self.view_text)
-
-    def clear_view(self):
-        self.view_text = ""
-        self.view.setText("")
-
-
-class OptionsView(QWidget):
-    """Menu for actions on the data processed from the SD log files."""
-
-    def __init__(self, parent: QWidget, data_model: DataModelManager):
-        super(OptionsView, self).__init__(parent)
-
-        self.data_model = data_model
-        self.data_model.layoutChanged.connect(self.modelUpdated)
-
-        header = QHBoxLayout()
-        header_label = QLabel("Data Options")
-        header_label.setStyleSheet("font-size: 24px; font-weight: 600")
-        header_info = NERInfoButton(SD_OPTIONS_DESC)
-        header.addStretch()
-        header.addWidget(header_label)
-        header.addWidget(header_info)
-        header.addStretch()
-
-        # Setup layouts
-        layout1: QVBoxLayout = self.generateLayout("Current Data")
-        layout2: QVBoxLayout = self.generateLayout("Filters")
-        layout3: QVBoxLayout = self.generateLayout("Export")
-
-        # Populate layout 1
-        self.model_info = QLabel("0 data points")
-        self.clear_button = NERButton("Clear", NERButton.Styles.RED)
-        self.clear_button.pressed.connect(self.clearModel)
-        self.clear_button.setToolTip("Delete all data in the application")
-        layout1.addWidget(self.model_info)
-        layout1.addWidget(self.clear_button)
-
-        # Populate layout 2
-        self.id_input = QLineEdit()
-        self.id_input.setToolTip("Space separated list of data IDs.")
-        self.filter_method = QCheckBox()
-        self.filter_method.setToolTip(
-            "Whether to keep or delete the specified IDs")
-        self.is_method_keep = True
-        self.filter_method.setChecked(self.is_method_keep)
-        self.setFilterMethod(self.is_method_keep)
-        self.filter_method.stateChanged.connect(self.setFilterMethod)
-        filter_layout = QGridLayout()
-        filter_layout.addWidget(QLabel("IDs:"), 0, 0)
-        filter_layout.addWidget(self.id_input, 0, 1)
-        filter_layout.addWidget(QLabel("Filter Method:"), 1, 0)
-        filter_layout.addWidget(self.filter_method, 1, 1)
-        self.filter_button = NERButton("Filter", NERButton.Styles.BLUE)
-        self.filter_button.pressed.connect(self.applyFilters)
-        self.filter_button.setToolTip("Start the filter")
-        layout2.addLayout(filter_layout)
-        layout2.addWidget(self.filter_button)
-
-        # Populate layout 3
-        self.graph_button = NERButton("Graph", NERButton.Styles.BLUE)
-        self.graph_button.setFixedWidth(150)
-        self.graph_button.setToolTip(
-            "Open a graph dashboard with the loaded data")
-        self.graph_button.pressed.connect(
-            lambda: GraphDialog(
-                self, self.data_model).exec())
-        self.csv_button = NERButton("CSV", NERButton.Styles.BLUE)
-        self.csv_button.setFixedWidth(150)
-        self.csv_button.setToolTip("Export the data to a CSV file")
-        self.csv_spinner = NERLoadingSpinner()
-        self.csv_button.pressed.connect(lambda: ExportDialog(
-            self, self.data_model, self.csv_spinner).exec())
-        csv_layout = QHBoxLayout()
-        csv_layout.addWidget(self.csv_button)
-        csv_layout.addWidget(self.csv_spinner)
-        self.database_button = NERButton("Database", NERButton.Styles.BLUE)
-        self.database_button.setFixedWidth(150)
-        self.database_button.setToolTip("Export the data to the database")
-        self.database_spinner = NERLoadingSpinner()
-        self.database_button.pressed.connect(lambda: DatabaseDialog(
-            self, self.data_model, self.database_spinner).exec())
-        database_layout = QHBoxLayout()
-        database_layout.addWidget(self.database_button)
-        database_layout.addWidget(self.database_spinner)
-        layout3.addWidget(self.graph_button)
-        layout3.addLayout(csv_layout)
-        layout3.addLayout(database_layout)
-
-        # Populate main layout
-        sub_layout = QHBoxLayout()
-        sub_layout.addStretch()
-        sub_layout.addLayout(layout1)
-        sub_layout.addStretch()
-        sub_layout.addLayout(layout2)
-        sub_layout.addStretch()
-        sub_layout.addLayout(layout3)
-        sub_layout.addStretch()
-        main_layout = QVBoxLayout()
-        main_layout.addLayout(header)
-        main_layout.addLayout(sub_layout)
-        self.setLayout(main_layout)
-
-    def modelUpdated(self):
-        self.model_info.setText(
-            f"{self.data_model.getDataCount()} data points")
-
-    def generateLayout(self, label: str):
-        layout = QVBoxLayout()
-        layout_label = QLabel(label)
-        layout_label.setStyleSheet("font-size: 16px; font-weight: 400")
-        layout_label.setAlignment(
-            Qt.AlignmentFlag.AlignHCenter | Qt.AlignmentFlag.AlignTop)
-        layout.addWidget(layout_label)
-        return layout
-
-    def clearModel(self):
-        self.data_model.deleteAllData()
-        self.modelUpdated()
-
-    def setFilterMethod(self, checked):
-        self.is_method_keep = checked
-        if self.is_method_keep:
-            self.filter_method.setText("(keep ids)")
-        else:
-            self.filter_method.setText("(remove ids)")
-
-    def applyFilters(self):
-        try:
-            filter_list = [int(id) for id in self.id_input.text().split(" ")]
-        except BaseException:
-            QMessageBox.critical(
-                self,
-                "ID Format Error",
-                "Expecting a list of IDs seperated by spaces.")
-            return
-
-        self.data_model.filter(filter_list, self.is_method_keep)
-
+from PyQt6.QtCore import QSize
 
 class SdCardWindow(QMainWindow):
     """Main window in for the SD Card connection."""
