+++ ner_telhub/view/vehicle/data_view/data_view.py
@@ -1,9 +1,8 @@
 from PyQt6.QtWidgets import (
     QWidget, QVBoxLayout
 )
-
 from ner_telhub.model.data_models import DataModelManager
-from ner_telhub.widgets.graphing_widgets import GraphDashboardWidget
+from ner_telhub.widgets.graphing_widgets.graphing_widgets import GraphDashboardWidget
 
 
 class DataView(QWidget):
