+++ ner_telhub/view/vehicle/map_view/date_time_encoder.py
@@ -1,0 +1,9 @@
+import datetime
+import json
+
+
+class DateTimeEncoder(json.JSONEncoder):
+    def default(self, obj):
+        if isinstance(obj, datetime):
+            return obj.isoformat()
+        return super(DateTimeEncoder, self).default(obj)
