+++ ner_telhub/widgets/graphing_widgets/format.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class Format(Enum):
+    """
+    Specifies available graph formats.
+    """
+    LINE = 1
