+++ ner_telhub/widgets/graphing_widgets/graph_dashboard.py
@@ -1,0 +1,7 @@
+from abc import abstractmethod
+from PyQt6.QtWidgets import QWidget
+
+class GraphDashboard(QWidget):
+    @abstractmethod
+    def removeGraph(self, graph):
+        pass
