+++ ner_telhub/widgets/styled_widgets/styled_widgets.py
@@ -1,0 +1,17 @@
+from enum import Enum
+import os
+
+from PyQt6.QtGui import QIcon, QMovie
+from PyQt6.QtWidgets import (
+    QHBoxLayout, QPushButton, QSizePolicy,
+    QToolBar, QWidget, QLabel, QMessageBox
+)
+
+
+
+
+
+
+
+
+
